Do not request decl for generic names
Summary:
Previously, there was a bug such that we were requesting the decls for generic names. Such as in the following code:

```
function foo<reify T>(T $_): void {...}
```

We were requesting the decl for `T`. We were doing this because of the following rationale:

For each function parameter, we check the type constraint of the parameter:

```
                match has_type_constraint(env, Option::from(param.type_info.as_ref()), ast_param) {
                    (L::Unconstrained, _) => Ok(instr::empty()),
                    (L::Not, _) => Ok(instr::verify_param_type(param_local)),
                    (L::Maybe, Some(h)) => {
                    (L::Definitely, Some(h)) => {
```

In the last two cases (where there was maybe a reified generic and where there was definitely a reified generic in the type constraint) we went a step furthered a checked the toplevel type parameter name for it's decl.

If it was a class with no reified generics, we would emit the simpler VerifyParam instruction. Otherwise we would emit the more complex VerifyParamTS.

This made sense if the following case were to occur:

```
class Box<T> {}

function foo<reify T>(Box<T> $_): void {}
```

In this case, even though `T` is a reified type parameter, because `Box` has only erased generics, we only need to check the `VerifyParam` instruction.

But the bug lies in the fact that the type parameter could just be reified generic, which we should not look up at all. This even produces an incorrect answer in the case where the name of the parameter is also the name of a real class that we can look up the decl for:

```
class T {}

function foo<reify T>(T $_): void {}
```

In this case, even though we should emit the `VerifyParamTS` instruction, we will look up `T`, find it to be a class without reified generics, and emit the wrong `VerifyParam` instruction.

This diff checks to ensure that the id that we are considering is not a generic before requesting the decl. If the id is a reified generic, we emit the `VerifyParamTS` instruction, otherwise if it a erased generic, emit the `VerifyParam` instruction, otherwise, look up the decl.

Reviewed By: edwinsmith

Differential Revision: D38555076

fbshipit-source-id: e4c45a628b2cd4d95880bb6d64577bb9c971ce11