Remove typechecker-only options from GlobalOptions to avoid code duplication
Summary:
Typechecker-only flags in `GlobalOptions` are prefixed with `tco_`.
However, they are used (inconsistently) through `GlobalOptions`,
which requires 2x definitions (`.ml` and `.mli` files)
and also defeats the purpose of why `typecheckerOptions` module exist.

This is the first step toward better encapsulation of `*Options`.
Later, we'd want to hide the representation of a set of related options in all but a single module
so that there is only a *single way* of reading/setting each option.
Currently, one can access *any* option through *any* of the following:
- `globalOptions`
- `parserOptions`
- `typecheckerOptions`
- `gleanOptions`
- `declParserOptions`
- `symbolWriteOptions`

Reviewed By: vassilmladenov

Differential Revision: D38281484

fbshipit-source-id: 09831124813528415a92f64c3ad62c02b9789df0