Change Test Parameters
Summary:
A number of tests currently turn off logging entirely.  This could instead be
more narrowly address by disabling external extractors.  This worked for all
cases but one.

The remaining case was throwing an exception (which turning off logging hid)
due to not having computed the hash of the file.  This now computes the hash
and passes it in to avoid the exception all together.

Reviewed By: rodmk

Differential Revision: D38602943

fbshipit-source-id: a27a2d300801187393e55e129b9b2ed5e21dbe46