Detailed SQLite db logging
Summary:
It can be really difficult to debug why accessing dbs is failing.  This adds a
ton of detailed logging around when that occurs to make it more easy to
diagnose.

The detailed logging is a little expensive, but it's only occuring when an
exception is being thrown and when facts isn't working, so the cost is well
justified by the debugging information in that situation.

Reviewed By: jthemphill

Differential Revision: D38582538

fbshipit-source-id: 825394a9be09b6acd92ebf7652202b91a2eb1cb8