Optimize ReflectionClass::getConstants()
Summary:
For very large class hierarchies, particularly those with many traits that implement the same interfaces this method can be excessively slow.

It's unclear to me why it needs to iterate all of the traits, interfaces, and parents to get an ordering, but my best guess is "PHP".

The `m_constants` field has a defined ordering based on where constants are pulled from that seems to be parents, then included enums, then declared interfaces, and then finally constants declared on the class itself. Probably not worth the effort to fix anyone depending on this weird existing ordering though.

Reviewed By: ricklavoie

Differential Revision: D38648322

fbshipit-source-id: 8982b991ca6abc143d7aae11f5f7b19b3b7541b4