Return `Optional<Type>` from jit::irgen::param_target_type
Summary: `param_target_type` uses `TBottom` as a sentinel value to indicate that a function parameter for a builtin does not need to be checked, storing such in `ParamPrep` as well. Instead of having two places we store this info, we have `param_target_type` return an `Optional<Type>`. We guard on this when we would have checked `ParamPrep::needsConversion` before and use an appropriate default elsewhere (generally `TUninit`).

Reviewed By: ricklavoie

Differential Revision: D38360688

fbshipit-source-id: 08432b5ecfe1e5ce259ee81cd500e75cf5df817c