Use indices in MethTabEntry instead of pointers
Summary:
This one was a bit of a pain. MethTabEntry stores a pointer to the
php::Func that it refers to. Like other cases, this is problematic
because soon we may not necessarily have that php::Func in memory
(php::Funcs are owned by their enclosing classes, which may not be
loaded). A more stable encoding is to instead use the class' name, and
the index into that class' method table.

This required a few changes to trait flattening, to make sure the
indices are kept consistent. At one point during the final pass, we
might delete some empty special functions. This deletion can cause the
indices to shift. Instead of removing from the method table, just
delete the func (and have other code skip over null pointers). Instead
of having res::Func store the MethTabEntry, have it store the
php::Func directly (it only ever used the func anyways).

Reviewed By: jano

Differential Revision: D38687199

fbshipit-source-id: 47a218e021384a5359b18ea27a526891e6f6927b