handle namespacing of primitives
Summary:
Bless a list of primitives so readonly checks pass in Hackc.

Just like in `parser/readonly_check.rs`, used in prod, we allow returning a value as some_primitive_type as mutable even when the value is readonly. Easiest to explain via the snapshot test: `primitives_namespacing.php` in D38462806.

## Why:

The readonly checking in `parser/readonly_check.rs` expects primitives to look like this: `int`.

But in hackc, we are seeing types after `elaborate_namespace_visitor`, which changes `int` to `\HH\int`.

Reviewed By: aorenste

Differential Revision: D38579242

fbshipit-source-id: 491c53de71e8849301497b88027bfff115f09b37