Calculate join points after the fact
Summary:
Previously, we had an explicit join constraint that represents the places in the dataflow graph where dataflow converges.

This has a number of problems:
1. It is redundant since the information can be recovered by analysing the graph with subset constraints. This makes it tedious as well as error prone because such convergence can still occur without an explicit join constraint.
2. Join points involving only two branches is limiting in two ways. First, there might be multiple incoming edges. It is not ideal to create tree structure to handle this. Second, the incoming edges might be generated from distinct parts of the AST, e.g., multiple returns. These are difficult to handle when the join requires both branches.

To address both of these problems, we let the solver infer the join points using an adjacency map.

This diff breaks more test cases than it fixes precisely because of (1). I forgot to generate join points where there is one in loops. That combined with the unrestricted propagation of optional keys (even if they are not optional) makes some inference results seem worse, but this is due to an orthogonal problem (T125888579).

Differential Revision: D38391353

fbshipit-source-id: 7fe6afd389e358ddb7f55dc5ad3051870f0d4567