Implement FromOcamlRep for hackrs decls
Summary:
Make it possible to directly convert an OCaml value to a hackrs decl, instead of converting to an oxidized_by_ref value, then converting that to a hackrs type.

This way, we lose the guarantee of correctness we get from the OCaml->oxidized_by_ref conversion, but we gain a substantial runtime benefit, particularly for reads of values stored in sharedmem in OCaml-slab format.

Reviewed By: shayne-fletcher

Differential Revision: D38678418

fbshipit-source-id: e326872ab556b7489044aa036a17f7aee6118181