hh_file_prefetcher
Summary:
A binary that fetches files from the remote file store (hack_sub1m_file_store/tree/v0.0.1/files) based on hash(file) + sha1(content), and stores the content in sharedmem.

Note that this is a "prove of concept" binary to iterate upon and which does not yet interface with the other warmup modules.

Reviewed By: bobrenjc93

Differential Revision: D38663620

fbshipit-source-id: 52dad85ad069ec1d5becd6c2ae6855978663b25c