Patch a rewrite in the direct decl parser
Summary:
In P513353812 Leo tested refinements in class-level
constraints and exposed a bug. It turns out that there
is a sketchy post processing phase in the direct decl
parser that was not extended when type refinements were
introduced. In this diff I update it and make Leo's example
work.

En passant, I also fixed unions and intersections that
happen to be denotable using an experimental feature.

Leo's example is added as a new typechecker test.

To prevent further screwups, a wildcard was turned into
an explicit list of cases (some of which, I believe, are not
possible and now panic).

Reviewed By: losvald

Differential Revision: D38746351

fbshipit-source-id: 7ea4d5c12fb0a15c0058753a8245c62d0aba8cff