Disable external facts extraction in unit tests
Summary:
With facts enabled, external fact extractors can be used.  These extractors may
try to make network connections, may emit output, or do a variety of other
things which would generate output and potentially cause tests to fail.  This
continues on the work of the diff that added a flag to address this by adding a
configuration binding and disabling external facts via config.ini so each
individual test doesn't need to specify this.

This only sets the value in the slow test set right now, it may need to be
expanded as NF gets used more and more.

Reviewed By: ricklavoie

Differential Revision: D38628584

fbshipit-source-id: 535a95d1eb937776e36f29404ee4282b33d93cba