Only have one type of attribute struct
Summary: I borrowed from serde the pattern of defining separate Container/Field/Variant structs. But they have so much in common that I think it's probably fine to collapse them into one type. The errors they emit will differ, but we can handle that with the AttrKind added in this diff. This does mean that the Attrs we parse for a Field will have a useless `prefix` field, but as long as we (eventually) emit an error for a `prefix` attribute in that position I think that's fine.

Reviewed By: edwinsmith

Differential Revision: D38774315

fbshipit-source-id: 0a3d4486a0dacb2f1110f434991705d8c8e011a2