Do not change code or position for error when checking method in dynamic mode
Summary:
When checking a function or method under dynamic assumptions, we wrap any typing error with a description "Function f is not dynamically callable", with a special error code. Unfortunately, this means that we don't honour FIXMEs in the body of the function that are silencing an error that is produced on both static and dynamic passes through the code.

For now, let's remove this use of the special error code, and preserve the underlying code. We might want to revisit error handling at some point e.g. somehow determine that an error is *specific* to the dynamic pass.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D38828460

fbshipit-source-id: 5c3361246c12587f4b0b4235fe2a3977b3bdf9e6