Simplify C++ File API
Summary:
`struct File` declares a public virtual `close` and implements a protected non-virtual `closeImpl`. structs derived from it then either 1) implement both (but in all cases their `close` just calls their `closeImpl`) 2) implement just `close` or 3) if they're doubly derived some of them implement neither and rely on their parent's implementation(s). In cases 1 and 2 whatever the derived struct implements calls `File::closeImpl` as part of it's implementation.

We don't need this complexity. Make the current `File::closeImpl` be the default implementation of `File::close`. Case 1 children that override both can have their implementations merged as well; case 2 and 3 children don't need to change API. Code that was calling a `closeImpl` can now call the corresponding `close` (and if it was an unqualified `closeImpl()` call we add the struct name so that we resolve to the same implementation, not the leafmost one).

`Socket::close` was a little special: `Socket` wasn't implementing `closeImpl`, and its `close` implementation just called `Socket::closeImpl` (i.e. the one inherited from `File`). None of the structs derived from it implemented either (case 3). This means we can delete `Socket::close` and have this whole branch of the hierarchy directly go to the implementation on `File`.

Finally, since we're here and we don't intend to expand this hierarchy any time soon, mark all non-overridden implementations of `close` as final. This leaves just the default implementation on `File` and the overriding one on `PlainFile` (that's overridden by `TempFile`) as non-final implementations.

After this diff, all the methods named `closeImpl` are on things extending `FileData` (and the totally unrelated `LdapLink::closeImpl`), whereas things extending `File` only have methods named `close`. This makes the code slightly less confusing to read.

Reviewed By: ricklavoie

Differential Revision: D38837099

fbshipit-source-id: 1f82018bcdcc90c138a5757d1fd8a7c9344f9561