Replace s_pcloseRet RDS_LOCAL with optional File::close out param
Summary: `s_pcloseRet` only exists so that pclose, which is basically just a call to `Pipe::close`, can return the exit status of the process at the other end of the pipe in addition to the boolean success/failure that all the implementations of the `File::close` API return. We don't need to use an `RDS_LOCAL` for this, we can give the `File::close` API an optional out param that only Pipe::close will write to and only the pclose implementation will read.

Reviewed By: ricklavoie

Differential Revision: D38837097

fbshipit-source-id: 7a0bf3a0d7312635a259096e89906d95d6dfb679