Don't autocomplete optional parameters
Summary:
Previously, when autocompleting a method, we would always provide a template for all parameters.

```
$foo->someMetho

// After accepting an autocomplete option
$foo->someMethod($a, $b, $c);
```

However, if `$b` and `$c` are optional parameters, this is inconvenient. The user is forced to delete the inserted text to get `->someMethod($a)`.

Instead, only insert required parameters when the user accepts an autocomplete result.

Reviewed By: hgoldstein

Differential Revision: D38768155

fbshipit-source-id: 1e73b8a08ecdd22c0d4394d4250b8ef18459b7c9