Remove close constraint tests
Summary:
Close constraint tests dump the constraint solver output and performs fixpoint iteration on it while combining definitions. This is useful for debugging but all the functionality of the test is subsumed by the `solve_constraint` style tests that output the end result. These tests a sensitive to simple changes in the walker and the solver so they lead to a lot of churn.

This diff removes them in favour of more functional constraints.

Differential Revision: D38827316

fbshipit-source-id: bfc15eb7b0787a0f02d1cad227f4690e08a26a18