deal with disable tests in old thrift runtime
Summary:
some tests under `thrift/facebook/java/swift` are disabled. Some of these tests are enable-able.

`testReconnectingWhenRpcResourcesIsShutdown` is sadly still not passing due to reason mentioned in the todo, for this case, remove `Test` to make sure it is not discoverable as a test

remove `testSchedulerRejection` since similar tests exists in `thrift/lib/java` and that test is passing

Differential Revision: D38879893

fbshipit-source-id: 3cfc62842035063cfe566178aefe42998ea7cc54