Pass the receiver position when checking function calls
Summary:
When type checking function calls, pass the receiver position as well as the outer expression position. Use this position when checking capabilities, as the arguments aren't relevant and can make it harder to see what's happening.

For function calls `foo(123)`, use `foo` as the receiver position. For method calls `$foo->bar(123)`, use `bar` as the receiver position.

Reviewed By: andrewjkennedy

Differential Revision: D38848044

fbshipit-source-id: 6a36707d2299b44b1c484ee87a2547009bc7db0f