hardcode parse-on-demand=true
Summary:
hhvm --hphp (aka hphpc) supports a `--parse-on-demand` flag, default true, that controls
a feature called "on demand parsing": files which are in the autoload
universe (but not excluded by CLI options) are included in the build if
they are transitively referenced by the original set of input files (via SymbolRefs).

Maybe a better name would be "compile-time autoloading". Anyway, the flag is
effectively a no-op if no autoload configuration is provided, since the only files
eligible to be transitively referenced are the ones in the initial input set.

No tests or production use cases disable this mode, and autoloading in general
has been a Hack feature for a long time. The false case is not even tested.
Lets remove the flag to reduce complexity.

Reviewed By: ricklavoie

Differential Revision: D38794211

fbshipit-source-id: 5e17f7a9597597bfbb8f866ce68055efefd5d181