UnitIndex should not exclude files when used for Decls
Summary:
UnitIndex intended to replace autoload_map.hdf, which was a full index of every
file in the Autoload.Query repo config; the autoload map did not exclude files
from the `hhvm --hphp` command line exclusion patterns.

For the purposes of `-vEval.EnableDecl=true` the intent is to make all source files
available, not just the source files that are eligible to be included in the build.
Otherwise, if EnableDecl==false, it's ok to filter files because they will get filtered
anyway in resolveOnDemand().

Reviewed By: ricklavoie

Differential Revision: D38794891

fbshipit-source-id: 34c70e10df01d62f5fa42e9c72285467bb5ca60c