Kill maybeCyclic check in SetNewElemVec
Summary:
If we are inserting a vec into itself, it will have a refcount of at least
two. One for the existence of the base and another for the stack value being
inserted. We don't need the extra third refcount, the COW check will handle
this properly and use the SetNewElemVec C++ helper. After the operation
the original value will still exist, so it is legitimate to incref it.

Reviewed By: ricklavoie

Differential Revision: D38896332

fbshipit-source-id: f547c1f6447853c2189c01bb96e08b5ea7105245