Add struct log logging to HHBBC passes, improve logging output
Summary:
Add infrastructure to log information about all HHBBC passes using a
StructuredLogEntry. Add support for logging using StructuredLogEntry
even when running HHBBC in stand-alone mode, or when HHBBC is
disabled.

Add automatic logging of extern_worker::Client stats to
StructuredLogEntry, as well as in the TRACE output.

Use folly::prettyPrint in a few places to improve the readability of
the log output (and tweak the output format a bit).

Reviewed By: edwinsmith

Differential Revision: D38899993

fbshipit-source-id: 5d02aa72c8f1bcc6e1f9dac017533c9417a4be53