Quit if compile fails
Summary:
Using `hhvm_wrapper.php -c` causes two separate commands to be invoked. When the first (compiling the repo) fails, it would be helpful for the script to immediately quit so that it's more obvious it failed in the first step.

When printing the command, we can still continue, since although the first has been run, we won't be running the second (since the expectation is for the user to probably run the first again along with the second?).

Reviewed By: ricklavoie

Differential Revision: D38909575

fbshipit-source-id: 0d1274253acdeb2d765af40b9792355730fc17a3