Make serde order for unordered containers explicit in BlobEncoder
Summary:
BlobEncoder will allow you to encode unordered contains. To (try to)
maintain a consistent serialization, BlobEncoder will first sort the
container's keys, then emit the contents in that order. Right now it
just uses std::sort for this purpose. The danger in this, however, is
that std::sort may not sort the keys in a sufficiently deterministic
order! For standard types (like std::string), this is fine, but if you
use StringData* as a key (for example), it will sort it by the pointer
addresses, which isn't deterministic at all.

Fix this by making the sort comparator explicit. If you wish to
serialization an unordered container, you must provide an ordering
functor which determines how the keys are to be ordered. You also have
to provide the functor for decoding, but it is ignored there (it is
needed to keep the interfaces symmetrical).

Reviewed By: edwinsmith

Differential Revision: D38920030

fbshipit-source-id: ca23b57fe9618b6feb2878b8c3adf887c7d1930a