Add multithreaded benchmarks for TDigest and QuantileEstimator
Summary:
Add a benchmark for adding values to a multithreaded TDigest using a
DigestBuilder.

This makes it easier to compare the performance against other algorithms, such
as the CPUShardedQuantileHistogram.

This benchmark is designed to mimic the performance of TDigest as used via
DigestBuider, BufferedStat, and QuantileEstimator. As a consequence, the buffer
space used is rather huge -- 0.5 MB for a 64-core machine.

Reviewed By: Gownta

Differential Revision: D38880483

fbshipit-source-id: 1b858ff891bfa3a522a6107edbd0ce2e21b252cd