Fix valueAt<T>() for NOBITS sections
Summary:
`ElfFile::valueAt()` returns a reference to the contents of the file a the specified file address for the memory mapped ELF file. This works fine for content that is present in the file (on disk) but does not work for sections like like `.bss` which are not in the file but rather zero-initialized at run time.

**.bss**
*Uninitialized data that contribute to the program's memory image. By definition, the system initializes the data with zeros when the program begins to run. The section occupies no file space, as indicated by the section type SHT_NOBITS.*

[[link](https://docs.oracle.com/cd/E19683-01/816-1386/6m7qcoblj/index.html#chapter6-73445)]

This diff checks for section type `SHT_NOBITS` and returns a reference to zero-initalized data from `valueAt<T>()` for those sections.

Reviewed By: jordalgo, jwiepert

Differential Revision: D38868825

fbshipit-source-id: 0a6eb3366949028a66d8a353bec8886e10c34bc8