Use static analysis to convert AnnotAction::ObjectCheck to AnnotAction::Pass
Summary:
If all possible types return AnnotAction::Pass, the type checking logic is
entirely bypassed. Otherwise a CFG with a bunch of CheckTypes and their
corresponding implementation is generated.

In the case of AnnotAction::ObjectCheck, we almost always know that the type
will pass and emit emptish block in that CFG. We then rely on DCE, simplify and
other CFG related optimization passes to reduce all that CFG to nothing.
Unfortunately it sometimes does not work as intended, resulting in a bloated
noop CFG that is full of phis.

While it is definitely worth fixing these optimizations and their weird
interactions, a CFG that is not emitted does not need to be optimized away.

Add a static analysis that converts ObjectCheck to Pass and bypass the CFG if
all other cases Pass too.

Reviewed By: ricklavoie

Differential Revision: D38931172

fbshipit-source-id: 81872eb7dcf98cd3a0a395e0d52393baa20d4c35