Do not unnecessarily constrain values in verifyPropType()
Summary:
The result of type verification is usually the same verified value, unless
it was a class or lazyclass coerced to a string. The verifyPropType() logic is
currently incapable[*] of using a phi to merge these outcomes into a single
SSATmp value. To workaround this limitation, it was constraining inputs to
DataTypeSpecific and if it was impossible to produce a guard, a fallback
mechanism was used.

This logic was unnecessarily applied even if coercion was not possible. Don't
do that, constrain the value only if the TC may coerce.

After this diff, a guard is needed only for typehints that involve strings.
Multiple translations won't be emitted anymore for typehints such as
`?int` and `?Foo`.

[*] Work in progress, need to fix various CFG optimizations that interact in a
weird way.

Reviewed By: ricklavoie

Differential Revision: D38931374

fbshipit-source-id: 4714e5a0165295c4eca3cfc698d52a0d0859a848