Add symbolic links to shared paths for dependencies (#9158)
Summary:
ShipIt does not support synchronize a fbsource path to multiple GitHub paths. As a workaround, D38917635 and D38925702 filtered out the common fbsource path for dependencies and synchronize the common path to a single GitHub path. This diff will add symbolic links to shared paths for first party dependencies.

Pull Request resolved: https://github.com/facebook/hhvm/pull/9158

Reviewed By: alexeyt

Differential Revision: D38918651

fbshipit-source-id: af57187b208ba2ee8d5411640715cf8a6c936032