Simplify early stack overflow check determination logic
Summary:
Do not take credit for the number of actually passed arguments, use the minimum
number of required parameters instead.

The regular stack overflow and surprise check will be moved from DV initializer
func entry to the Enter opcode, where the number of originally passed arguments
is no longer available.

Most functions have zero optional arguments and won't be impacted. For other
functions, the number of optional arguments is usually way less than 256 (the
number of TVs in a page), so the impact is negligible.

Reviewed By: ricklavoie

Differential Revision: D38954127

fbshipit-source-id: 4dc7cf1f402b9461c549672a3eedb84d220d72cb