Subtyping support for expression-dependent types
Summary:
This diff teaches the typechecker to recognize that
runtime values have a class type with all type
constants defined. We may not know what they
are precisely if our static type is too loose, and so
we may retort to making a fresh unknown type
(potentially with bounds attached). The logic to
make these fresh types is compatible with what
Typing_taccess does, and is added to the
Typing_type_member module.

For simplicity of the implementation in the
subtyping engine I added a new constraint type
called Thas_type_member. This helps separating
refinements from their class type and limits
redundant work in Typing_subtype.

Reviewed By: CatherineGasnier

Differential Revision: D38787112

fbshipit-source-id: bef17fcb710e972ec847574bbe652d0f978bc9c2