Quick clarification between Class_get and Class_const in comment
Summary:
After playing around with this with Wilfred, we realized that there were some mistakes in the comments of these two AAST nodes. This diff corrects them. Namely, we determined that `Foo::bar()` is not actually a `Class_get` but is actually represented as a `Class_const`. As far as we can tell, `Class_get` indicates where the left hand side of the `::` is a value that starts with `$` - based on the logic in `lowerer.rs` here

https://www.internalfb.com/code/fbsource/[e5fa9ee7f6211383d5c5ccb10ab85ba3e7466565]/fbcode/hphp/hack/src/parser/lowerer/lowerer.rs?lines=2002-2013

I've updated the comments to match

Reviewed By: aorenste, Wilfred

Differential Revision: D38969213

fbshipit-source-id: eecfba41322545b8f84429b149aa7cea8b6f552a