Offer a naming table based shallow decl provider for hackc testing
Summary:
Currently, when testing decl directed bytecode in hackc independently of hhvm we only can request decls in the file being compiled. This diff adds an option to give a naming table and the path to the root of the naming table to hackc to allow it to look into more files for decls.

The two new options are

```
--naming-table ~/naming_table.nt --naming-table-root $REP_ROOT
```

These two options only work if hackc is invoked with ` --test-compile-with-decls`

Because of how we construct naming tables, the filepath associated with a symbol name is relative to a root directory, but it doesn't appear that we can remove the full path just from the naming table alone, so we ask that you pass in the root of the naming table as well.

In addition to using the naming table, hackc will also return decls from HHIs.

One can create a naming table with internal Meta tooling

Reviewed By: mheiber

Differential Revision: D38852201

fbshipit-source-id: 9fc42d4c853f4e20b4b5df9f647dad90c1986953