Avoid thundering herds of failing fallback jobs in extern-worker
Summary:
If there's a bug in a job you're running, then most (all?) of the RE
workers will crash. We'll automatically attempt to fallback to local
execution and re-run the jobs. Since the jobs are bugged, these will
crash the same way. So you get an enormous amount of log spam and
potentially 1000s of crashing local workers (each coredumping). This
isn't very useful.

Fix this by controlling the number of allowed concurrent fallback
executions with a semaphore. The semaphore's count is initially 1. We
must acquire the semaphore before running a fallback execution. When
the execution finishes (whether it is successful or throws), we
release the token back to the semaphore. So, only one execution is
allowed simultaneously. *However* if the execution finishes
successfully, we'll release an additional token back to the
semaphore. This increases the number of concurrent fallbacks
allowed. So, if the fallbacks are all crashing, the 1st one will
(probably) crash, we'll report that, and the rest won't get a chance
to run. If we fellback for some transient reason, however, the
fallbacks will succeed, and we'll quickly ramp up to full amount
allowed.

If we have a number of outstanding coroutines (say from within
coro::collectRange), and one throws, the rest will get
"cancelled". This was manifesting as a RE error. Add logic to detect
this, and manually check for cancellations at certain safe points. A
cancellation just silently kills the operation.

One last tweak: If the worker returns a non-zero exit code, we'll use
a different exception than just Error. This lets callers distinguish
that case. This is useful when using optimistic mode. We don't want to
retry the action if the worker is failing.

Reviewed By: edwinsmith

Differential Revision: D39007756

fbshipit-source-id: ffb1ec766781e9ed572832b0a3a965f49a04723f