Update READMEs
Summary:
Remove most of the content of `thrift/compiler/README` because it is either inaccurate or redundant (or both). In particular, we shouldn't duplicate build instructions because building only the compiler is just a matter of passing an additional CMake option.

Instead clarify which dependencies apply to compiler in the main README, add missing dependencies and list CMake options.

Reviewed By: avalonalex

Differential Revision: D39002982

fbshipit-source-id: d55f08fe488e5d8fbbbf0b69e5faebd68e60ad25