Optional field as what is not a definitive field
Summary:
Rather than using optional field constraints directly, we reformulate finding of optional fields as marking what is not definite as optiomal.

The nice thing about this formulation is that if a key is definitely in the dictionary, its status persists as opposed to further optional looking accesses poisoning the result.

Differential Revision: D38708175

fbshipit-source-id: e867a21d4299c7856724a241fb18a09d64a1ec4d