Refactor ternary operator handling
Summary:
There was unnecessary duplication between the case the then expression exists and the case it doesn't, i.e., `cond ? then : else` and `cond ?: else`.

This diff makes the common code shared between the two.

It also fixes a bug where in the `?:` case we were unconditionally evaluating the else branch.

Differential Revision: D38829974

fbshipit-source-id: a37858837e971a682660d44a63b9207396354bff