Refresh the environment while branching
Summary:
When we exit a conditional, we generate a join point, but when we enter it we use the same environment.

This was causing inaccurate results because if there is a `$d['a']` in one branch, we generated a `Has` constraint on the current entity which was shared between both branches making it as if `'a'` is guaranteed to exist in both branches.

This diff fixes it by refreshing the environment each time we enter a new branch.

Differential Revision: D38827315

fbshipit-source-id: 45a1b00ed6cfea3318a56dddb98caf2d00536c6a