Log codemod directives during full init
Summary:
Writes codemod directives to files of the form `/tmp/shape-like-dict-N.json` where `N` is replaced with the worker id creating it. This identifier is 1 to ~20 in case of an invocation via `hh` and 0 when it is invoked with `hh_single_type_check`.

Log level 3 is reserved for this logging mode. The minimum invocation is using

```
hh --config log_levels='{"shape_analysis":3}' --config defer_class_declaration_threshold=999999
```

Reviewed By: viratyosin

Differential Revision: D38880843

fbshipit-source-id: bb0bef9af195ced9d5fd002b04e0f2eb1c837e25