Separate static key constraints that provide info vs query
Summary:
A copy-on-write assignment provide information about going forward in the control-flow graph, e.g., "shape $d has field 'a' with type `int` going forward".

We need a version of this that says "shape $d **needs** to have a field 'a' with type `int`". This is useful in figuring out fields for parameters (as we can propagate the constraint backwards rather than forwards).

Differential Revision: D38827314

fbshipit-source-id: a90ede168638e965147c3cf545a3d9c13a9d9054