Refactor monitor so that we exit when hitting consecutive exceptions
Summary:
We have this loop in the monitor where we'll loop. If we hit exceptions in the loop that we can stomach and proceed through, we'll continue the loop. There is code that is supposed to count the number of consecutive exceptions and then exit once we hit some threshold (currently 500 consecutive exceptions). That code is here:

https://www.internalfb.com/code/fbsource/[9b1343c6c1b2755f11e2de9acf675300ef466c95]/fbcode/hphp/hack/src/monitor/serverMonitor.ml?lines=833-834%2C847-860

However, the problem with this code is that I'm fairly certain that this code path never gets triggered. That is because the code that it is wrapping in a try-catch loop does it's own error swallowing and doesn't send exceptions back up to the top loop.

The code that silently swallows the exceptions in the function is here:

https://www.internalfb.com/code/fbsource/[9b1343c6c1b2755f11e2de9acf675300ef466c95]/fbcode/hphp/hack/src/monitor/serverMonitor.ml?lines=864%2C894-896%2C900%2C906-907%2C911%2C917

This diff attempts to bubble the exceptions back up to the toplevel loop that is keeping track of consecutive exceptions. This way, in the future, if a new exception comes along and spins forever, we'll at least exit the monitor.

Reviewed By: CatherineGasnier

Differential Revision: D38965035

fbshipit-source-id: 3e64fca1ec0215e3e33455e0b59d9b43eb91fb43